Move metadata tag assignment actions into module
